refactor: re-organize ignitor code
