+++ src/Ignitor/ErrorHandler/index.ts
